requirements.txt
Flask
python-dotenv
langchain
langchain-community
langchain-google-genai
unstructured
faiss-cpu
